rename: post -> spot으로 이름 변경 design: 디자인 적용되지 않은 부분 다시 적용
