--- frontend/src/app/_components/post/index.ts
@@ -1,4 +1,0 @@
-import Post from './Post';
-import PostList from './PostList';
-
-export { Post, PostList };
