+++ frontend/src/app/_components/spot/index.ts
@@ -1,0 +1,4 @@
+import Spot from './Spot';
+import SpotList from './SpotList';
+
+export { Spot, SpotList };
